requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.25.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
httpx>=0.27.0
asgi-lifespan>=2.1.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
passlib>=1.7.4
argon2-cffi>=23.1.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
cloudinary>=1.36.0
